Moved remaining node packages into brambl layer since brambl depends on them. Also removed files that are no longer needed: Genus package (not used for these examples), node/package.scala.
