+++ src/main/scala/co/topl/brambl/wallet/CredentiallerError.scala
@@ -1,7 +1,7 @@
 package co.topl.brambl.wallet
 
 import co.topl.brambl.models.transaction.Attestation
-import co.topl.node.transaction.authorization
+import co.topl.brambl.authorization.{ValidationError => AuthorizationValidationError}
 
 sealed abstract class CredentiallerError
 
@@ -9,6 +9,6 @@ object CredentiallerErrors {
 
   abstract class ProverError extends CredentiallerError
   case class AttestationMalformed(attestation: Attestation) extends ProverError
-  case class ValidationError(error: authorization.ValidationError) extends CredentiallerError
+  case class ValidationError(error: AuthorizationValidationError) extends CredentiallerError
 
 }
