--- src/main/scala/co/topl/genus/OutputState.scala
@@ -1,8 +1,0 @@
-package co.topl.genus
-
-sealed abstract class OutputState
-
-object OutputStates {
-  case object Spent extends OutputState
-  case object Unspent extends OutputState
-}
