--- src/main/scala/co/topl/genus/TransactionOutput.scala
@@ -1,8 +1,0 @@
-package co.topl.genus
-
-//case class TransactionOutput(
-//  id:       Box.Id,
-//  boxValue: Box.Value,
-//  address:  Address,
-//  state:    OutputState = OutputStates.Unspent
-//)
