--- src/main/scala/co/topl/node/package.scala
@@ -1,6 +1,0 @@
-package co.topl
-
-package object node {
-  type SmallData = Array[Byte] // small, up to 64 bytes
-  type Root = Array[Byte] // fixed size 32 or 64 bytes as a root from an accumulator
-}
